+++ python/python/tests/test_async_client.py
@@ -39,6 +39,7 @@ from glide.async_commands.core import (
     InfBound,
     InfoSection,
     InsertPosition,
+    OnlyIfEqual,
     UpdateOptions,
 )
 from glide.async_commands.sorted_set import (
@@ -452,6 +453,7 @@ class TestCommands:
     async def test_conditional_set(self, glide_client: TGlideClient):
         key = get_random_string(10)
         value = get_random_string(10)
+
         res = await glide_client.set(
             key, value, conditional_set=ConditionalChange.ONLY_IF_EXISTS
         )
@@ -466,6 +468,29 @@ class TestCommands:
         )
         assert res is None
         assert await glide_client.get(key) == value.encode()
+        # Tests for ONLY_IF_EQUAL below in test_set_only_if_equal()
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    @pytest.mark.skip_if_version_below("8.1.0")
+    async def test_set_only_if_equal(self, glide_client: TGlideClient):
+        key = get_random_string(10)
+        value = get_random_string(10)
+        value2 = get_random_string(10)
+        wrong_comparison_value = get_random_string(10)
+        while wrong_comparison_value == value:
+            wrong_comparison_value = get_random_string(10)
+
+        await glide_client.set(key, value)
+
+        res = await glide_client.set(
+            key, "foobar", conditional_set=OnlyIfEqual(wrong_comparison_value)
+        )
+        assert res is None
+        assert await glide_client.get(key) == value.encode()
+        res = await glide_client.set(key, value2, conditional_set=OnlyIfEqual(value))
+        assert res == OK
+        assert await glide_client.get(key) == value2.encode()
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])